// src/app/enums.ts
export enum ClothingSize {
    XS = 'XS',
    S = 'S',
    M = 'M',
    L = 'L',
    XL = 'XL',
    XXL = 'XXL'
}

export enum OrderStatus {
    Delivered = 'DELIVERED',
    InProgress = 'IN_PROGRESS',
    Canceled = 'CANCELED',
}

export enum Gender {
    Male = 'male',
    Female = 'female',
}

export enum ClothingType {
    SportsWear = 'sports wear',
    Casual = 'casual',
    Official = 'official'
}